Move BaseUserCore to caliopen_main.common.core location
