+++ src/backend/main/py.main/caliopen_main/common/core/__init__.py
@@ -2,7 +2,8 @@
 """Caliopen common core classes."""
 from __future__ import absolute_import, print_function, unicode_literals
 
+from .base import BaseUserCore
 from .pubkey import PublicKey
 from .related import BaseUserRelatedCore
 
-__all__ = ['PublicKey', 'BaseUserRelatedCore']
+__all__ = ['BaseUserCore', 'PublicKey', 'BaseUserRelatedCore']
