+++ src/backend/main/py.storage/caliopen_storage/core/__init__.py
@@ -1,9 +1,6 @@
 from .registry import core_registry
-
-from .base import BaseCore, BaseUserCore
+from .base import BaseCore
 from .mixin import MixinCoreRelation
 
 __all__ = [
-    'core_registry',
-    'BaseCore', 'BaseUserCore',
-    'MixinCoreRelation']
+    'core_registry', 'BaseCore', 'MixinCoreRelation']
